曲選択画面まで

+++ Make_Dialog_MusicInfo.h
@@ -10,16 +10,16 @@
 
 namespace Make {
 	namespace Dialog {
-		constexpr const std::uint8_t NAMECHARMAX{50};
-		constexpr const std::uint8_t ARTISTCHARMAX{50};
-		constexpr const std::uint8_t LEVELCHARMAX{2};
-		constexpr const std::uint8_t BPMCHARMAX{6};
-		constexpr const std::uint8_t TOTALMINUTESCHARMAX{4};
-		constexpr const std::uint8_t BEGINDELAYCHARMAX{5};
-		constexpr const std::uint8_t LEVELMAX{ 20 };
-		constexpr const std::uint8_t LEVELMIN{ 1 };
-		constexpr const std::uint8_t LANEMAX{8};
-		constexpr const std::uint8_t LANEMIN{4};
+		constexpr const std::uint16_t NAMECHARMAX{50};
+		constexpr const std::uint16_t ARTISTCHARMAX{50};
+		constexpr const std::uint16_t LEVELCHARMAX{2};
+		constexpr const std::uint16_t BPMCHARMAX{6};
+		constexpr const std::uint16_t TOTALMINUTESCHARMAX{4};
+		constexpr const std::uint16_t BEGINDELAYCHARMAX{5};
+		constexpr const std::uint16_t LEVELMAX{ 20 };
+		constexpr const std::uint16_t LEVELMIN{ 1 };
+		constexpr const std::uint16_t LANEMAX{8};
+		constexpr const std::uint16_t LANEMIN{4};
 		class Make_Dialog_MusicInfo
 		{
 		private:
@@ -27,7 +27,7 @@ namespace Make {
 			static bool isInputed;
 			static INT_PTR CALLBACK MusicInfoDialogProc(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
 		public:
-			void getMusicInfoFromDlg(char(&filePath)[MAX_PATH], char(&name)[MAX_PATH],char(&artist)[MAX_PATH],std::uint8_t& level,double& bpm, double& totalMinutes, double& beginDelay);
+			void getMusicInfoFromDlg(char(&filePath)[MAX_PATH], char(&name)[MAX_PATH],char(&artist)[MAX_PATH],std::uint16_t& level,double& bpm, double& totalMinutes, double& beginDelay);
 		};
 	}
 }
