+++ Make_Note_NormalNoteContainer.h
@@ -12,9 +12,9 @@ namespace Make {
 		private:
 			static double noteWidth; //描画するノーツの幅；
 			const std::uint16_t barID; //何小節目に属しているか
-			const std::uint8_t beatID;//その小節の何番目の線か
+			const std::uint16_t beatID;//その小節の何番目の線か
 			const double time;//曲の開始から何秒か
-			const std::uint8_t laneAmount; //レーンの数
+			const std::uint16_t laneAmount; //レーンの数
 			const double& r_y; //拍線の座標
 			std::uint32_t color;
 			std::vector<bool> noteFlag;
@@ -22,13 +22,13 @@ namespace Make {
 			double notePointX;
 			double notePointY;
 		public:
-			Make_Note_NormalNoteContainer(const std::uint16_t barID,const std::uint8_t beatID,const double& y,const std::uint8_t amountOfLane,const double time);
-			void setNormalNoteFlag(const std::uint8_t laneID);
+			Make_Note_NormalNoteContainer(const std::uint16_t barID,const std::uint16_t beatID,const double& y,const std::uint16_t amountOfLane,const double time);
+			void setNormalNoteFlag(const std::uint16_t laneID);
 			void drawNote();
-			const bool getNormalNoteFlag(const std::uint8_t laneID);
+			const bool getNormalNoteFlag(const std::uint16_t laneID);
 			const double& getTime();
 			const std::uint16_t& getBarID();
-			const std::uint8_t& getBeatID();
+			const std::uint16_t& getBeatID();
 		};
 	}
 }
